+++ CHANGES.md
@@ -5,7 +5,7 @@
 
 ### New features & changes
 * Added support for attributed string properties (`attributedText` in `UILabel`, `UITextField` and `UITextView`, and `attributedTitle` in `UIButton`).
-* Added support for opting out of scheduled styling when adding style classes.
+* Added support for opting out of scheduled styling when adding and removing style classes (via `UIView (InterfaCSS)` category).
 
 ### Bug fixes
 * Fixed problem with using multiple constants within the same property value.